# openapi/filmorate.yml
openapi: 3.0.3
info:
  title: Filmorate
  description: Приложение для обмена мнением о фильмах
  version: 1.0.0

servers:
  - url: /filmorate

paths:
  /films:
    get:
      operationId: getFilms
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Film'
        '500':
          description: На любое исключение на сервере
    post:
      operationId: createFilm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Film'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Film'
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: На любое исключение на сервере
    put:
      operationId: updateFilm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Film'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Film'
        '405':
          description: Invalid input
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: На любое исключение на сервере

  /films/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      operationId: getFilmById
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Film'
        '404':
          description: Not found
        '405':
          description: Invalid input
        '400':
          description: Bad request
        '500':
          description: На любое исключение на сервере

  /films/{id}/like/{userId}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    put:
      operationId: addLike
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: На любое исключение на сервере
    delete:
      operationId: deleteLike
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: На любое исключение на сервере

  /films/popular:
    get:
      operationId: getPopular
      parameters:
        - name: count
          in: query
          required: false
          description: Если count = 0, то вернется первые 10 фильмов
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Film'
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: На любое исключение на сервере

  /users:
    get:
      operationId: getUsers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: На любое исключение на сервере
    post:
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '405':
          description: Invalid input
        '400':
          description: Bad request
        '500':
          description: На любое исключение на сервере
    put:
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '405':
          description: Invalid input
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: На любое исключение на сервере

  /users/{id}:
    get:
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: На любое исключение на сервере

  /users/{id}/friends/{friendId}:
    put:
      operationId: addFriend
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: friendId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: boolean
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: На любое исключение на сервере
    delete:
      operationId: deleteFriend
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: friendId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: boolean
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: На любое исключение на сервере

  /users/{id}/friends:
    get:
      operationId: getFriends
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: На любое исключение на сервере


  /users/{id}/friends/common/{otherId}:
    get:
      operationId: getCommonFriends
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: otherId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: На любое исключение на сервере

  /mpa/{id}:
    get:
      operationId: getMpaNameById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MPA'
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: На любое исключение на сервере
  /mpa:
    get:
      operationId: getAll
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MPA'
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: На любое исключение на сервере

  /genres/{id}:
    get:
      operationId: getGenreById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Genre'
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: На любое исключение на сервере
  /genres:
    get:
      operationId: getAllGenres
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Genre'
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: На любое исключение на сервере

components:
  schemas:
    Film:
      type: object
      required:
        - name
        - description
        - releaseDate
        - duration
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          nullable: false
          pattern: ^(?!\s*$).+
        description:
          type: string
          maxLength: 200
        releaseDate:
          type: string
          description: "Дата релиза не раньше 28.12.1895"
          example: 1900-12-28
        duration:
          type: integer
          format: int64
          minimum: 1
        mpa:
          $ref: '#/components/schemas/MPA'
          description: "Возрастной рейтинг фильма"
          example: PG-13
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'
          description: "Жанр фильма"
          example: Комедия
        likes:
          type: array
          items:
            type: integer
            format: int64
            uniqueItems: true

    MPA:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          enum:
            - "G"
            - "PG"
            - "PG-13"
            - "R"
            - "NC-17"
    Genre:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        nameEng:
          type: string
          enum:
            - "COMEDY"
            - "DRAMA"
            - "CARTOON"
            - "THRILLER"
            - "DOCUMENTARY"
            - "ACTION"
    User:
      type: object
      required:
        - email
        - login
        - birthday
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          nullable: false
          pattern: ^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$
        login:
          type: string
          pattern: ^[^\s]*\S[^\s]*$
        name:
          type: string
          default: null
          description: Если имя = null, то используем в качестве него логин
        birthday:
          type: string
          description: Дата рождения не может быть в будущем
          example: 1946-08-20
        friends:
          type: array
          items:
            type: integer
            format: int64
            uniqueItems: true